Pass validator arguments through for string guard
